♻️ F09: Rebased handleCardDelete and reloadMenu behaviours
